Conflicts:
	src/main/java/com/mycompany/csc229_211review_lab02hw/Student.java

+++ src/main/java/com/mycompany/csc229_211review_lab02hw/DriverClass.java
@@ -7,13 +7,14 @@ package com.mycompany.csc229_211review_lab02hw;
 public class DriverClass {
 	
 	public static void main(String[] args) {
+            
 		// ToDo 5: Fix the error
 		
 		// ToDo 6: Fix the constructor of Student class
 		
 		// ToDo 7: Add a toString method for Student class
 		
-		Student std1= new Student("James", 20);
+		Student std1= new Student("James", (short) 20);
 		
 		// ToDo 8: Set the gpa of the student using the scanner and user
 		// 			input and then print the output.
@@ -24,5 +25,9 @@ public class DriverClass {
 		
 		// ToDo 10: submit using a pull request.
 	}
-
 }
+
+
+
+        
+
